+++ exporter/elasticsearchexporter/config_test.go
@@ -36,19 +36,20 @@ func TestLoadConfig(t *testing.T) {
 	require.NoError(t, err)
 	require.NotNil(t, cfg)
 
-	assert.Equal(t, len(cfg.Exporters), 2)
+	assert.Equal(t, len(cfg.Exporters), 3)
 
 	defaultCfg := factory.CreateDefaultConfig()
 	defaultCfg.(*Config).Endpoints = []string{"https://elastic.example.com:9200"}
 	r0 := cfg.Exporters[config.NewComponentID(typeStr)]
 	assert.Equal(t, r0, defaultCfg)
 
-	r1 := cfg.Exporters[config.NewComponentIDWithName(typeStr, "customname")].(*Config)
+	r1 := cfg.Exporters[config.NewComponentIDWithName(typeStr, "trace")].(*Config)
 	assert.Equal(t, r1, &Config{
-		ExporterSettings: config.NewExporterSettings(config.NewComponentIDWithName(typeStr, "customname")),
+		ExporterSettings: config.NewExporterSettings(config.NewComponentIDWithName(typeStr, "trace")),
 		Endpoints:        []string{"https://elastic.example.com:9200"},
 		CloudID:          "TRNMxjXlNJEt",
-		Index:            "myindex",
+		LogsIndex:        "logs-generic-default",
+		TracesIndex:      "trace_index",
 		Pipeline:         "mypipeline",
 		HTTPClientSettings: HTTPClientSettings{
 			Authentication: AuthenticationSettings{
@@ -79,6 +80,45 @@ func TestLoadConfig(t *testing.T) {
 			Dedot: true,
 		},
 	})
+
+	r2 := cfg.Exporters[config.NewComponentIDWithName(typeStr, "log")].(*Config)
+	assert.Equal(t, r2, &Config{
+		ExporterSettings: config.NewExporterSettings(config.NewComponentIDWithName(typeStr, "log")),
+		Endpoints:        []string{"http://localhost:9200"},
+		CloudID:          "TRNMxjXlNJEt",
+		LogsIndex:        "my_log_index",
+		TracesIndex:      "traces-generic-default",
+		Pipeline:         "mypipeline",
+		HTTPClientSettings: HTTPClientSettings{
+			Authentication: AuthenticationSettings{
+				User:     "elastic",
+				Password: "search",
+				APIKey:   "AvFsEiPs==",
+			},
+			Timeout: 2 * time.Minute,
+			Headers: map[string]string{
+				"myheader": "test",
+			},
+		},
+		Discovery: DiscoverySettings{
+			OnStart: true,
+		},
+		Flush: FlushSettings{
+			Bytes: 10485760,
+		},
+		Retry: RetrySettings{
+			Enabled:         true,
+			MaxRequests:     5,
+			InitialInterval: 100 * time.Millisecond,
+			MaxInterval:     1 * time.Minute,
+		},
+		Mapping: MappingsSettings{
+			Mode:  "ecs",
+			Dedup: true,
+			Dedot: true,
+		},
+	})
+
 }
 
 func withDefaultConfig(fns ...func(*Config)) *Config {