+++ exporter/elasticsearchexporter/model.go
@@ -16,15 +16,17 @@ package elasticsearchexporter // import "github.com/open-telemetry/opentelemetry
 
 import (
 	"bytes"
-
+	"encoding/json"
 	"go.opentelemetry.io/collector/pdata/pcommon"
 	"go.opentelemetry.io/collector/pdata/plog"
+	"go.opentelemetry.io/collector/pdata/ptrace"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/elasticsearchexporter/internal/objmodel"
 )
 
 type mappingModel interface {
 	encodeLog(pcommon.Resource, plog.LogRecord) ([]byte, error)
+	encodeSpan(pcommon.Resource, ptrace.Span) ([]byte, error)
 }
 
 // encodeModel tries to keep the event as close to the original open telemetry semantics as is.
@@ -48,6 +50,32 @@ func (m *encodeModel) encodeLog(resource pcommon.Resource, record plog.LogRecord
 	document.AddInt("SeverityNumber", int64(record.SeverityNumber()))
 	document.AddAttribute("Body", record.Body())
 	document.AddAttributes("Attributes", record.Attributes())
+	document.AddAttributes("Resource", resource.Attributes())
+
+	if m.dedup {
+		document.Dedup()
+	} else if m.dedot {
+		document.Sort()
+	}
+
+	var buf bytes.Buffer
+	err := document.Serialize(&buf, m.dedot)
+	return buf.Bytes(), err
+}
+
+func (m *encodeModel) encodeSpan(resource pcommon.Resource, span ptrace.Span) ([]byte, error) {
+	var document objmodel.Document
+	document.AddTimestamp("@timestamp", span.StartTimestamp()) // We use @timestamp in order to ensure that we can index if the default data stream logs template is used.
+	document.AddTimestamp("endTimestamp", span.EndTimestamp()) // We use @timestamp in order to ensure that we can index if the default data stream logs template is used.
+	document.AddID("TraceId", span.TraceID())
+	document.AddID("SpanId", span.SpanID())
+	document.AddID("ParentSpanId", span.ParentSpanID())
+	document.AddString("Name", span.Name())
+	document.AddString("Kind", span.Kind().String())
+	document.AddInt("TraceStatus", int64(span.Status().Code()))
+	linkJson, _ := json.Marshal(span.Links())
+	document.AddString("Link", string(linkJson))
+	document.AddAttributes("Attributes", span.Attributes())
 	document.AddAttributes("Resource", resource.Attributes())
 
 	if m.dedup {