+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - `signalfxexporter`: Enable the exporting of seven Kubernetes metrics used in Splunk/SignalFx content by default (#11032)
 - `googlecloudexporter`: Support writing to multiple GCP projects by setting the `gcp.project.id` resource attribute, and support service account impersonation (#11051)
 - `k8sattributeprocessor`: Add debug logs to help identify missing attributes (#11060)
+- `elasticsearchexporter`: Support sending OpenTelemetry tracing data to Elasticsearch index or data stream.(#10180)
 
 ### 🧰 Bug fixes 🧰